Add escaping in javascript used to build HTML
In addition to escaping the HTML strings that are being built on the
feedbacks page (This is the only page I could find building HTML out of
user-entered data), the links created on that page are also being cleaned
up. They are validated against a regex, and then processed from an
absolute link into a relative link so that no links on that page can be
exploited to trick a user into going somewhere outside the application.