XML collector tweaks to handle UT XML collection
Clear the current timestamp value buffer before appending the new value
and trying to parse it. With previous collections there has been a single
timestamp field, but with the UT collection we are getting one for every
obstype value and just using the last one processed.

Clear some collector member variables when an exception is caught.
May not come up as much in production environments, but the SSL
connection was commonly closed when debugging the UT collection.
Accumulated values were then causing any future collection attempts to
fail.

Add a '0' to the added values counter when the document is started.
During the processing of the close tag for the root element there was a
NoSuchElementException thrown during every processing (not just for UT,
but for all collections). That exception has been happening for all XML
collections, but it was being swallowed, and hasn't interfered with the
actual processing since it happens at the very end.